Providing better basic examples, cleaning
Fixed some small grammar mistakes. Fixed an error in the Table of Contents and headers. Added yarn for install. Simplified and expanded the Basic example to show a full working document, and provided the resulting code for clarity. Added more variables to the Dynamic Values example, including showing passing a value as a prop for clarity, and provided the resulting code for clarity.